# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Swagger with Laravel",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/Passenger/create": {
            "get": {
                "tags": [
                    "Passenger"
                ],
                "summary": "Create 30 Fake Data in Passenger",
                "operationId": "a8d578d74a24af93b35275f9ffef16a7",
                "responses": {
                    "200": {
                        "description": "Create 30 Fake Data in Passenger"
                    },
                    "500": {
                        "description": "Server Error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/Passenger/store": {
            "post": {
                "tags": [
                    "Passenger"
                ],
                "summary": "Create a Passenger",
                "operationId": "95ea0a0313abb1f793238d6bf039491a",
                "parameters": [
                    {
                        "name": "firstName",
                        "in": "query",
                        "description": "firstName Passenger",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "lastName",
                        "in": "query",
                        "description": "lastName Passenger",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "nationalcode",
                        "in": "query",
                        "description": "nationalcode Passenger",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "gender",
                        "in": "query",
                        "description": "gender Passenger (male,femail)",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "birthday",
                        "in": "query",
                        "description": "birthday Passenger",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Passenger detail successfully"
                    },
                    "409": {
                        "description": "TrainTicket Validate Error"
                    },
                    "500": {
                        "description": "Server Error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/Passenger/show": {
            "get": {
                "tags": [
                    "Passenger"
                ],
                "summary": "User Passengers",
                "operationId": "ea322437a151b91877ba0072053b0e3a",
                "responses": {
                    "200": {
                        "description": "get Passengers User successfully"
                    },
                    "500": {
                        "description": "Server Error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/Passenger/edit": {
            "get": {
                "tags": [
                    "Passenger"
                ],
                "summary": "edit a Passenger",
                "operationId": "d6894fae42581a8f4ca6c98adb97f994",
                "parameters": [
                    {
                        "name": "id",
                        "in": "query",
                        "description": "id Passenger",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Passenger detail successfully"
                    },
                    "500": {
                        "description": "Server Error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/Passenger/update": {
            "put": {
                "tags": [
                    "Passenger"
                ],
                "summary": "Update a Passenger",
                "operationId": "0843b09d4174a93ff943ff1fb4cc95bc",
                "parameters": [
                    {
                        "name": "id",
                        "in": "query",
                        "description": "id Passenger",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "firstName",
                        "in": "query",
                        "description": "firstName Passenger",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "lastName",
                        "in": "query",
                        "description": "lastName Passenger",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "nationalcode",
                        "in": "query",
                        "description": "nationalcode Passenger",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "gender",
                        "in": "query",
                        "description": "gender Passenger (male,femail)",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "birthday",
                        "in": "query",
                        "description": "birthday Passenger",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Passenger detail successfully"
                    },
                    "409": {
                        "description": "TrainTicket Validate Error"
                    },
                    "500": {
                        "description": "Server Error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/Passenger/destroy": {
            "delete": {
                "tags": [
                    "Passenger"
                ],
                "summary": "Delete a Passenger",
                "operationId": "43178b44f79acb77fa193c4570a2d8f1",
                "parameters": [
                    {
                        "name": "id",
                        "in": "query",
                        "description": "id Passenger",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Passenger detail successfully"
                    },
                    "500": {
                        "description": "Server Error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/Passenger/restore": {
            "put": {
                "tags": [
                    "Passenger"
                ],
                "summary": "Restore a Passenger",
                "operationId": "13e8568f2fe4b1f473994502b6d7f009",
                "parameters": [
                    {
                        "name": "id",
                        "in": "query",
                        "description": "id Passenger",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Passenger detail successfully"
                    },
                    "500": {
                        "description": "Server Error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    }
}